requirements.txt
dnspython==2.7.0
duckdb==1.3.1
numpy==2.2.6
pandas==2.3.0
pymongo==4.13.2
pyspark==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
